+++ config.py
@@ -1,7 +1,17 @@
+from datetime import datetime
+
+debug = True
+
+port_to_run = 5000
+
 schedule_verification_time = "00:30"
+
+# In a new system when the schedule should start?
+firs_datetime = datetime(2016, 7, 2)
 
 path_to_database = './alph.db'
 
 secret_key = 'abrakadabra'
 
-
+node_manager_address = '127.0.0.1'
+node_manager_port = '45064'
